+++ builds/wtf_wikipedia.mjs
@@ -1,4 +1,4 @@
-/* wtf_wikipedia 8.3.0 MIT */
+/* wtf_wikipedia 8.4.0 MIT */
 import https from 'https';
 
 var parseUrl = function parseUrl(url) {
@@ -70,7 +70,7 @@ function _unsupportedIterableToArray(o, minLen) {
   if (typeof o === "string") return _arrayLikeToArray(o, minLen);
   var n = Object.prototype.toString.call(o).slice(8, -1);
   if (n === "Object" && o.constructor) n = o.constructor.name;
-  if (n === "Map" || n === "Set") return Array.from(n);
+  if (n === "Map" || n === "Set") return Array.from(o);
   if (n === "Arguments" || /^(?:Ui|I)nt(?:8|16|32)(?:Clamped)?Array$/.test(n)) return _arrayLikeToArray(o, minLen);
 }
 
@@ -89,7 +89,7 @@ function _nonIterableRest() {
 var isInterWiki = /(wiktionary|wikinews|wikibooks|wikiquote|wikisource|wikispecies|wikiversity|wikivoyage|wikipedia|wikimedia|foundation|meta)\.org/;
 var defaults = {
   action: 'query',
-  prop: 'revisions',
+  prop: 'revisions|pageprops',
   //we use the 'revisions' api here, instead of the Raw api, for its CORS-rules..
   rvprop: 'content',
   maxlag: 5,
@@ -181,10 +181,20 @@ var getResult = function getResult(data, options) {
       text = page.revisions[0].slots.main['*'];
     }
 
+    page.pageprops = page.pageprops || {};
+    var domain = options.domain;
+
+    if (!domain && options.wiki) {
+      domain = "".concat(options.wiki, ".org");
+    }
+
     var meta = Object.assign({}, options, {
       title: page.title,
       pageID: page.pageid,
-      namespace: page.ns
+      namespace: page.ns,
+      domain: domain,
+      wikidata: page.pageprops.wikibase_item,
+      description: page.pageprops['wikibase-shortdesc']
     });
 
     try {
@@ -549,7 +559,7 @@ var toJson$1 = function toJson(img, options) {
 
 var toJson_1 = toJson$1;
 
-var server = 'https://wikipedia.org/wiki/Special:Redirect/file/';
+var server = 'wikipedia.org';
 
 var encodeTitle = function encodeTitle(file) {
   var title = file.replace(/^(image|file?)\:/i, ''); //titlecase it
@@ -600,12 +610,15 @@ var methods = {
     return [];
   },
   url: function url() {
-    return server + makeSrc(this.file());
+    // let lang = 'en' //this.language() || 'en' //hmm: get actual language?
+    var fileName = makeSrc(this.file());
+    var domain = this.data.domain || server;
+    var path = "wiki/Special:Redirect/file";
+    return "https://".concat(domain, "/").concat(path, "/").concat(fileName);
   },
   thumbnail: function thumbnail(size) {
     size = size || 300;
-    var path = makeSrc(this.file());
-    return server + path + '?width=' + size;
+    return this.url() + '?width=' + size;
   },
   format: function format() {
     var arr = this.file().split('.');
@@ -674,6 +687,20 @@ var methods$1 = {
 
     return this.data.pageID;
   },
+  wikidata: function wikidata(id) {
+    if (id !== undefined) {
+      this.data.wikidata = id;
+    }
+
+    return this.data.wikidata;
+  },
+  domain: function domain(str) {
+    if (str !== undefined) {
+      this.data.domain = str;
+    }
+
+    return this.data.domain;
+  },
   language: function language(lang) {
     if (lang !== undefined) {
       this.data.lang = lang;
@@ -689,7 +716,7 @@ var methods$1 = {
     }
 
     var lang = this.language() || 'en';
-    var domain = this.data.domain || 'wikipedia.org'; // replace blank to underscore
+    var domain = this.domain() || 'wikipedia.org'; // replace blank to underscore
 
     title = title.replace(/ /g, '_');
     title = encodeURIComponent(title);
@@ -776,6 +803,8 @@ var methods$1 = {
     return this.sentences(0);
   },
   images: function images(clue) {
+    var _this2 = this;
+
     var arr = _sectionMap(this, 'images', null); //grab image from infobox, first
 
     this.infoboxes().forEach(function (info) {
@@ -791,6 +820,8 @@ var methods$1 = {
         obj.images = obj.images || [];
         obj.images.forEach(function (img) {
           if (img instanceof Image_1 === false) {
+            img.language = _this2.language();
+            img.domain = _this2.domain();
             img = new Image_1(img);
           }
 
@@ -2167,7 +2198,7 @@ var interwiki$1 = parseInterwiki;
 
 var ignore_links = /^:?(category|catégorie|Kategorie|Categoría|Categoria|Categorie|Kategoria|تصنيف|image|file|image|fichier|datei|media):/i;
 var external_link = /\[(https?|news|ftp|mailto|gopher|irc)(:\/\/[^\]\| ]{4,1500})([\| ].*?)?\]/g;
-var link_reg = /\[\[(.{0,160}?)\]\]([a-z]+)?(\w{0,10})/gi; //allow dangling suffixes - "[[flanders]]s"
+var link_reg = /\[\[(.{0,160}?)\]\]([a-z]+)?/gi; //allow dangling suffixes - "[[flanders]]s"
 
 var external_links = function external_links(links, str) {
   str.replace(external_link, function (raw, protocol, link, text) {
@@ -2319,11 +2350,17 @@ var kill_xml = function kill_xml(wiki) {
   //only kill ref tags if they are selfclosing
 
   wiki = wiki.replace(/ ?< ?(ref) [a-zA-Z0-9=" ]{2,100}\/ ?> ?/g, ' '); //<ref name="asd"/>
-  //some formatting xml, we'll keep their insides though
+  // convert these html tags to known formatting
+
+  wiki = wiki.replace(/<i>(.*?)<\/i>/g, "''$1''");
+  wiki = wiki.replace(/<b>(.*?)<\/b>/g, "'''$1'''"); // these are better-handled with templates
+
+  wiki = wiki.replace(/<sub>(.*?)<\/sub>/g, "{{sub|$1}}");
+  wiki = wiki.replace(/<sup>(.*?)<\/sup>/g, "{{sup|$1}}"); //some formatting xml, we'll keep their insides though
 
   wiki = wiki.replace(/ ?<[ \/]?(p|sub|sup|span|nowiki|div|table|br|tr|td|th|pre|pre2|hr)[ \/]?> ?/g, ' '); //<sub>, </sub>
 
-  wiki = wiki.replace(/ ?<[ \/]?(abbr|bdi|bdo|blockquote|cite|del|dfn|em|i|ins|kbd|mark|q|s|small)[ \/]?> ?/g, ' '); //<abbr>, </abbr>
+  wiki = wiki.replace(/ ?<[ \/]?(abbr|bdi|bdo|blockquote|cite|del|dfn|em|ins|kbd|mark|q|s|small)[ \/]?> ?/g, ' '); //<abbr>, </abbr>
 
   wiki = wiki.replace(/ ?<[ \/]?h[0-9][ \/]?> ?/g, ' '); //<h2>, </h2>
 
@@ -4350,7 +4387,7 @@ var imgLayouts = {
   "super": true
 }; //images are usually [[image:my_pic.jpg]]
 
-var oneImage = function oneImage(img) {
+var oneImage = function oneImage(img, doc) {
   var m = img.match(file_reg);
 
   if (m === null || !m[2]) {
@@ -4366,7 +4403,9 @@ var oneImage = function oneImage(img) {
 
   if (title) {
     var obj = {
-      file: file
+      file: file,
+      lang: doc.lang,
+      domain: doc.domain
     }; //try to grab other metadata, too
 
     img = img.replace(/^\[\[/, '');
@@ -4389,20 +4428,20 @@ var oneImage = function oneImage(img) {
       obj.caption = parseSentence$4(arr[arr.length - 1]);
     }
 
-    return new Image_1(obj, img);
+    return new Image_1(obj);
   }
 
   return null;
 };
 
-var parseImages = function parseImages(paragraph) {
+var parseImages = function parseImages(paragraph, doc) {
   var wiki = paragraph.wiki; //parse+remove scary '[[ [[]] ]]' stuff
 
   var matches = nested_find_1(wiki);
   matches.forEach(function (s) {
     if (isFile.test(s) === true) {
       paragraph.images = paragraph.images || [];
-      var img = oneImage(s);
+      var img = oneImage(s, doc);
 
       if (img) {
         paragraph.images.push(img);
@@ -4559,7 +4598,7 @@ var parse$5 = {
   list: list
 };
 
-var parseParagraphs = function parseParagraphs(section) {
+var parseParagraphs = function parseParagraphs(section, doc) {
   var wiki = section.wiki;
   var paragraphs = wiki.split(twoNewLines); //don't create empty paragraphs
 
@@ -4576,7 +4615,7 @@ var parseParagraphs = function parseParagraphs(section) {
 
     parse$5.list(paragraph); // parse images
 
-    parse$5.image(paragraph); //parse the sentences
+    parse$5.image(paragraph, doc); //parse the sentences
 
     parseSentences(paragraph);
     return new Paragraph_1(paragraph);
@@ -4614,6 +4653,7 @@ var normalize = function normalize(str) {
 
 var Infobox = function Infobox(obj) {
   this._type = obj.type;
+  this.domain = obj.domain;
   Object.defineProperty(this, 'data', {
     enumerable: false,
     value: obj.data
@@ -4658,6 +4698,8 @@ var methods$9 = {
     var obj = s.json();
     obj.file = obj.text;
     obj.text = '';
+    obj.domain = this.domain; // add domain information for image
+
     return new Image_1(obj);
   },
   get: function get() {
@@ -5718,6 +5760,18 @@ var templates = {
     }
 
     return str;
+  },
+  //https://en.wikipedia.org/wiki/Template:Sub
+  sub: function sub(tmpl, list) {
+    var obj = parse$3(tmpl, ['text']);
+    list.push(obj);
+    return obj.text || '';
+  },
+  //https://en.wikipedia.org/wiki/Template:Sup
+  sup: function sup(tmpl, list) {
+    var obj = parse$3(tmpl, ['text']);
+    list.push(obj);
+    return obj.text || '';
   }
 }; //aliases
 
@@ -6340,19 +6394,19 @@ var parsers$1 = {
     return '';
   },
   // https://en.wikipedia.org/wiki/Template:See
-  'see': function see(tmpl, list) {
+  see: function see(tmpl, list) {
     var obj = parse$3(tmpl);
     list.push(obj);
     return '';
   },
   // https://en.wikipedia.org/wiki/Template:For
-  'for': function _for(tmpl, list) {
+  "for": function _for(tmpl, list) {
     var obj = parse$3(tmpl);
     list.push(obj);
     return '';
   },
   // https://en.wikipedia.org/wiki/Template:Further
-  'further': function further(tmpl, list) {
+  further: function further(tmpl, list) {
     var obj = parse$3(tmpl);
     list.push(obj);
     return '';
@@ -6364,7 +6418,7 @@ var parsers$1 = {
     return '';
   },
   // https://en.wikipedia.org/wiki/Template:Listen
-  'listen': function listen(tmpl, list) {
+  listen: function listen(tmpl, list) {
     var obj = parse$3(tmpl);
     list.push(obj);
     return '';
@@ -6468,7 +6522,8 @@ var parsers$1 = {
     images = images.map(function (file) {
       var img = {
         file: file
-      };
+      }; // TODO: add lang and domain information
+
       return new Image_1(img).json();
     });
     obj = {
@@ -6598,7 +6653,7 @@ var all = {
 }; //a bunch of aliases for these ones:
 // https://en.wikipedia.org/wiki/Category:Tournament_bracket_templates
 
-var brackets = ['2teambracket', '4team2elimbracket', '8teambracket', '16teambracket', '32teambracket', 'cwsbracket', 'nhlbracket', 'nhlbracket-reseed', '4teambracket-nhl', '4teambracket-ncaa', '4teambracket-mma', '4teambracket-mlb', '8teambracket-nhl', '8teambracket-mlb', '8teambracket-ncaa', '8teambracket-afc', '8teambracket-afl', '8teambracket-tennis3', '8teambracket-tennis5', '16teambracket-nhl', '16teambracket-nhl divisional', '16teambracket-nhl-reseed', '16teambracket-nba', '16teambracket-swtc', '16teambracket-afc', '16teambracket-tennis3', '16teambracket-tennis5'];
+var brackets = ['2teambracket', '4team2elimbracket', '8teambracket', '16teambracket', '32teambracket', '4roundbracket-byes', 'cwsbracket', 'nhlbracket', 'nhlbracket-reseed', '4teambracket-nhl', '4teambracket-ncaa', '4teambracket-mma', '4teambracket-mlb', '16teambracket-two-reseeds', '8teambracket-nhl', '8teambracket-mlb', '8teambracket-ncaa', '8teambracket-afc', '8teambracket-afl', '8teambracket-tennis3', '8teambracket-tennis5', '16teambracket-nhl', '16teambracket-nhl divisional', '16teambracket-nhl-reseed', '16teambracket-nba', '16teambracket-swtc', '16teambracket-afc', '16teambracket-tennis3', '16teambracket-tennis5'];
 brackets.forEach(function (key) {
   all[key] = all['4teambracket'];
 });
@@ -8170,7 +8225,7 @@ var isInfobox$1 = function isInfobox(obj) {
 }; //reduce the scary recursive situations
 
 
-var allTemplates = function allTemplates(section) {
+var allTemplates = function allTemplates(section, doc) {
   var wiki = section.wiki; // nested data-structure of templates
 
   var list = find(wiki);
@@ -8215,6 +8270,8 @@ var allTemplates = function allTemplates(section) {
     }
 
     if (isInfobox$1(obj) === true) {
+      obj.domain = doc.domain; //
+
       section.infoboxes.push(new Infobox_1(obj));
       return false;
     }
@@ -8237,7 +8294,7 @@ var parseSentence$6 = _04Sentence.fromText; //okay, <gallery> is a xml-tag, with
 //all deities help us. truly -> https://en.wikipedia.org/wiki/Help:Gallery_tag
 // - not to be confused with https://en.wikipedia.org/wiki/Template:Gallery...
 
-var parseGallery = function parseGallery(section) {
+var parseGallery = function parseGallery(section, doc) {
   var wiki = section.wiki;
   wiki = wiki.replace(/<gallery([^>]*?)>([\s\S]+?)<\/gallery>/g, function (_, attrs, inside) {
     var images = inside.split(/\n/g);
@@ -8248,7 +8305,9 @@ var parseGallery = function parseGallery(section) {
     images = images.map(function (str) {
       var arr = str.split(/\|/);
       var obj = {
-        file: arr[0].trim()
+        file: arr[0].trim(),
+        lang: doc.language,
+        domain: doc.domain
       };
       var img = new Image_1(obj).json();
       var caption = arr.slice(1).join('|');
@@ -8458,9 +8517,9 @@ var math$1 = parseMath;
 // ... others are {{start}}...{{end}}
 // -> these are those ones.
 
-var xmlTemplates = function xmlTemplates(section) {
+var xmlTemplates = function xmlTemplates(section, doc) {
   election(section);
-  gallery(section);
+  gallery(section, doc);
   math$1(section);
   mlb(section);
   mma(section);
@@ -8482,11 +8541,11 @@ var parse$6 = {
 };
 
 var oneSection = function oneSection(section, doc) {
-  parse$6.startEndTemplates(section); //parse-out the <ref></ref> tags
+  parse$6.startEndTemplates(section, doc); //parse-out the <ref></ref> tags
 
   parse$6.references(section); //parse-out all {{templates}}
 
-  parse$6.templates(section); // //parse the tables
+  parse$6.templates(section, doc); // //parse the tables
 
   parse$6.table(section); //now parse all double-newlines
 
@@ -8776,7 +8835,7 @@ var fetchRandom = function fetchRandom(lang, options) {
     url = "https://".concat(options.domain, "/").concat(options.path, "?");
   }
 
-  url += "format=json&action=query&generator=random&grnnamespace=0&prop=revisions&rvprop=content&grnlimit=1&rvslots=main&origin=*";
+  url += "format=json&action=query&generator=random&grnnamespace=0&prop=revisions|pageprops&rvprop=content&grnlimit=1&rvslots=main&origin=*";
   var headers = _headers(options);
   return server$1(url, headers).then(function (res) {
     try {
@@ -8891,7 +8950,7 @@ var fetchCategory = function fetchCategory(category, lang, options) {
 
 var category = fetchCategory;
 
-var _version = '8.3.0';
+var _version = '8.4.0';
 
 var wtf = function wtf(wiki, options) {
   return _01Document(wiki, options);