+++ src/02-section/start-to-end/index.js
@@ -7,9 +7,9 @@ const parseMath = require('./math')
 // Most templates are '{{template}}', but then, some are '<template></template>'.
 // ... others are {{start}}...{{end}}
 // -> these are those ones.
-const xmlTemplates = function(section) {
+const xmlTemplates = function (section, doc) {
   parseElection(section)
-  parseGallery(section)
+  parseGallery(section, doc)
   parseMath(section)
   parseMlb(section)
   parseMMA(section)