+++ src/_fetch/02-getResult.js
@@ -1,8 +1,8 @@
 //this data-format from mediawiki api is nutso
-const getResult = function(data, options) {
+const getResult = function (data, options) {
   options = options || {}
   let pages = Object.keys(data.query.pages)
-  let docs = pages.map(id => {
+  let docs = pages.map((id) => {
     let page = data.query.pages[id] || {}
     if (page.hasOwnProperty('missing') || page.hasOwnProperty('invalid')) {
       return null
@@ -13,10 +13,18 @@ const getResult = function(data, options) {
     if (!text && page.revisions[0].slots) {
       text = page.revisions[0].slots.main['*']
     }
+    page.pageprops = page.pageprops || {}
+    let domain = options.domain
+    if (!domain && options.wiki) {
+      domain = `${options.wiki}.org`
+    }
     let meta = Object.assign({}, options, {
       title: page.title,
       pageID: page.pageid,
-      namespace: page.ns
+      namespace: page.ns,
+      domain: domain,
+      wikidata: page.pageprops.wikibase_item,
+      description: page.pageprops['wikibase-shortdesc'],
     })
     try {
       return { wiki: text, meta: meta }