+++ src/image/index.js
@@ -23,11 +23,11 @@ const imgLayouts = {
   baseline: true,
   middle: true,
   sub: true,
-  super: true
+  super: true,
 }
 
 //images are usually [[image:my_pic.jpg]]
-const oneImage = function(img) {
+const oneImage = function (img, doc) {
   let m = img.match(file_reg)
   if (m === null || !m[2]) {
     return null
@@ -40,7 +40,9 @@ const oneImage = function(img) {
   title = title.replace(/ /g, '_')
   if (title) {
     let obj = {
-      file: file
+      file: file,
+      lang: doc.lang,
+      domain: doc.domain,
     }
     //try to grab other metadata, too
     img = img.replace(/^\[\[/, '')
@@ -55,24 +57,23 @@ const oneImage = function(img) {
       obj.alt = imgData.alt
     }
     //remove 'thumb' and things
-    arr = arr.filter(str => imgLayouts.hasOwnProperty(str) === false)
+    arr = arr.filter((str) => imgLayouts.hasOwnProperty(str) === false)
     if (arr[arr.length - 1]) {
       obj.caption = parseSentence(arr[arr.length - 1])
     }
-    return new Image(obj, img)
+    return new Image(obj)
   }
   return null
 }
 
-const parseImages = function(paragraph) {
+const parseImages = function (paragraph, doc) {
   let wiki = paragraph.wiki
-
   //parse+remove scary '[[ [[]] ]]' stuff
   let matches = nested_find(wiki)
-  matches.forEach(function(s) {
+  matches.forEach(function (s) {
     if (isFile.test(s) === true) {
       paragraph.images = paragraph.images || []
-      let img = oneImage(s)
+      let img = oneImage(s, doc)
       if (img) {
         paragraph.images.push(img)
       }