+++ changelog.md
@@ -2,6 +2,15 @@
 
 #### [unreleased]
 
+#### 8.4.0
+
+- new `wikidata()` method
+- new `domain()` method
+- support image urls from 3rd-party wikis
+- support for some html formatting tags \#374
+- support for sub and sup templates
+- [fix] for link-parsing bug \#375
+
 ### 8.3.0
 
 - adds some wikivoyage templates